feat(linter): detect if eslint configs should be migrated to monorepo style
