+++ packages/linter/src/generators/init/init-migration.ts
@@ -1,0 +1,9 @@
+import type { Tree } from '@nrwl/devkit';
+
+export function migrateConfigToMonorepoStyle(
+  tree: Tree,
+  project: string
+): void {
+  // TODO: Make Check if this applies
+  console.log('This should migrate the monorepo');
+}
